chore: extract IngredientImage to be reusable
